+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Retour Changelog
 
+## 4.1.4 - 2022.10.03
+### Changed
+* Move to using `ServicesTrait` and add getter methods for services
+* Users must have permission to access Retour to be able to use the Retour widget ([#251](https://github.com/nystudio107/craft-retour/issues/251))
+
+### Fixed
+* Ensure that `$driver` is nullable in the install migration ([#252](https://github.com/nystudio107/craft-retour/issues/252)) ([#250](https://github.com/nystudio107/craft-retour/pull/250))
+
 ## 4.1.3 - 2022.09.14
 ### Changed
 * Fixed an issue where element types that update their `uriFormat` directly rather than re-saving the elements will generate automatic element redirects ([#248](https://github.com/nystudio107/craft-retour/issues/248))